feat: add resources filtering and sorting (#489)
This pull request restores the filtering and sorting functionality for the resource tables.

In the resources page (which is paginated on the server-side), filtering and sorting is also implemented server-side. For rest of resource tables (the resource children table and the project resources table), filtering and sorting is implemented in client-side. The table-related React components have been updated to support this hybrid behavior.

Additionally, a bunch of very small fixes have been applied as well.

Closes #481.