feat: make pagination smarter
The new cursor used to paginate events holds a page and an offset. These two are used to make sure that the list of events returned by default (which belongs to the latest events in the redis array) is always the size of a full page (given that there are equal or more items in the array to fill a full page).